removed commented out functions

+++ services/CommentingServiceImpl.php
@@ -21,16 +21,18 @@ class CommentingServiceImpl implements CommentingService {
 	 * @ReturnType Comment
 	 */
 	public function addComment(Comment $comment) {
-       // if(StudentServiceImpl::getInstance()->verifyAuth()) {
+       if(StudentServiceImpl::getInstance()->verifyAuth()) {
             $comment->setStudentid(StudentServiceImpl::getInstance()->getCurrentStudentId());
             $commentDAO = new CommentDAO();
             $commentDAO->create($comment);
 
-        //}
+        }
 	}
 	public function getLastInsertId(Comment $comment){
-        $commentDAO = new CommentDAO();
-	    return $commentDAO->getLastInsertId($comment);
+        if(StudentServiceImpl::getInstance()->verifyAuth()) {
+            $commentDAO = new CommentDAO();
+            return $commentDAO->getLastInsertId($comment);
+        }
     }
 
 	/**
@@ -41,14 +43,14 @@ class CommentingServiceImpl implements CommentingService {
 	 * @ReturnType Comment
 	 */
 	public function deleteComment(&$id) {
-        //if(StudentServiceImpl::getInstance()->verifyAuth()) {
+        if(StudentServiceImpl::getInstance()->verifyAuth()) {
             $commentDAO = new CommentDAO();
             $comment = $commentDAO->read($id);
             if($comment->getStudentid() == StudentServiceImpl::getInstance()->getCurrentStudentId())
             {
                 $commentDAO->delete($id);
             }
-	    //}
+	    }
     }
 	/**
 	 * @access public
@@ -59,7 +61,7 @@ class CommentingServiceImpl implements CommentingService {
 	 */
 	public function readCommentsForModule(&$moduleId)
     {
-        //if(StudentServiceImpl::getInstance()->verifyAuth()) {
+        if(StudentServiceImpl::getInstance()->verifyAuth()) {
             $commentDAO = new CommentDAO();
             $commentVoteDAO = new CommentVoteDAO();
             $studentDAO = new StudentDAO();
@@ -79,7 +81,7 @@ class CommentingServiceImpl implements CommentingService {
                 }
             }
             return $comments;
-      //  }
+       }
 	}
 
 	/**
@@ -90,14 +92,14 @@ class CommentingServiceImpl implements CommentingService {
 	 * @ReturnType Comment
 	 */
 	public function updateComment(Comment $comment) {
-        //if(StudentServiceImpl::getInstance()->verifyAuth()) {
+        if(StudentServiceImpl::getInstance()->verifyAuth()) {
             $commentDAO = new CommentDAO();
             $commentToEdit = $commentDAO->read($comment->getId());
             if($commentToEdit->getStudentid() == StudentServiceImpl::getInstance()->getCurrentStudentId())
             {
                 $commentDAO->update($comment);
             }
-       // }
+       }
 	}
 
 	/**
@@ -106,7 +108,7 @@ class CommentingServiceImpl implements CommentingService {
 	 * @ParamType id int
 	 */
 	public function voteOnComment(&$id) {
-      //  if(StudentServiceImpl::getInstance()->verifyAuth()) {
+      if(StudentServiceImpl::getInstance()->verifyAuth()) {
             $commentVoteDAO = new CommentVoteDAO();
             $studentid = StudentServiceImpl::getInstance()->getCurrentStudentId();
             $commentVote = new CommentVote();
@@ -114,7 +116,7 @@ class CommentingServiceImpl implements CommentingService {
             $commentVote->setStudentid($studentid);
             $commentVoteDAO->create($commentVote);
 
-       // }
+      }
 	}
     /**
      * @access public
@@ -122,12 +124,12 @@ class CommentingServiceImpl implements CommentingService {
      * @ParamType id int
      */
     public function deleteVote(&$id) {
-        //  if(StudentServiceImpl::getInstance()->verifyAuth()) {
-        $commentVoteDAO = new CommentVoteDAO();
-        $studentid = StudentServiceImpl::getInstance()->getCurrentStudentId();
-        $commentVoteDAO->delete($id,$studentid);
+        if(StudentServiceImpl::getInstance()->verifyAuth()) {
+            $commentVoteDAO = new CommentVoteDAO();
+            $studentid = StudentServiceImpl::getInstance()->getCurrentStudentId();
+            $commentVoteDAO->delete($id,$studentid);
 
-        // }
+        }
     }
 }
 ?>
